requirements.txt
# NEUROSENSE - Raspberry Pi 5 Sensor System
# Pin this to exact versions for stable deployment
RPi.bme280==0.2.4
smbus2==0.4.3
grove.py==0.6.0
numpy==1.26.4
flask==3.1.0
lgpio==0.2.2.0

# Development / testing only
pytest==8.3.4
